# docs/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/v1/catalog": {
      "get": {
        "operationId": "MastersController_list",
        "parameters": [
          {
            "name": "kind",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "Ítems de un catálogo (incluye inactivos; el selector filtra por active).",
        "tags": [
          "masters"
        ]
      },
      "post": {
        "operationId": "MastersController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CatalogDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "masters"
        ]
      }
    },
    "/api/v1/catalog/{id}": {
      "patch": {
        "operationId": "MastersController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CatalogUpdateDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "masters"
        ]
      }
    },
    "/api/v1/catalog/{id}/active": {
      "post": {
        "operationId": "MastersController_setActive",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ActiveDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "masters"
        ]
      }
    },
    "/api/v1/clients/lookup": {
      "get": {
        "operationId": "MastersController_lookup",
        "parameters": [
          {
            "name": "documentType",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "document",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "Cliente por documento (404 si es la primera vez).",
        "tags": [
          "masters"
        ]
      }
    },
...
    "/api/v1/alerts/{id}/{action}": {
      "post": {
        "operationId": "MonitoringController_alertAction",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "action",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AlertActionDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "acknowledge | manage | resolve | close (cierre de Alta/Crítica exige revisión senior).",
        "tags": [
          "monitoring"
        ]
      }
    },
    "/api/v1/telemetry/positions": {
      "post": {
        "operationId": "MonitoringController_ingest",
        "parameters": [
          {
            "name": "X-Integration-Key",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TelemetryBatchDto"
              }
            }
          }
        },
        "responses": {
          "202": {
            "description": ""
          }
        },
        "summary": "Recibe lotes de posiciones (≤ 500) del adaptador del proveedor GPS. Idempotente por vehículo + instante.",
        "tags": [
          "monitoring"
        ]
      }
    },
...
    "/api/v1/passenger-bookings": {
      "get": {
        "operationId": "PassengersController_list",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "passengers"
        ]
      },
      "post": {
        "operationId": "PassengersController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PassengerDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "passengers"
        ]
      }
    },
...
    "/api/v1/passenger-bookings/{id}": {
      "patch": {
        "operationId": "PassengersController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PassengerUpdateDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "passengers"
        ]
      }
    },
    "/api/v1/passenger-bookings/{id}/advance": {
      "post": {
        "operationId": "PassengersController_advance",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AdvanceDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "passengers"
        ]
      }
    },
    "/api/v1/trips/{id}/manifest": {
      "get": {
        "operationId": "PassengersController_manifest",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "passengers"
        ]
      }
    },
    "/api/v1/passengers/lookup": {
      "get": {
        "operationId": "PassengersController_lookup",
        "parameters": [
          {
            "name": "documentType",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "document",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "passengers"
        ]
      }
    },
...
    "/api/v1/me/bookings/{id}/cancel": {
      "post": {
        "operationId": "PassengersController_cancelMine",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CancelDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "passengers"
        ]
      }
    },
...
  },
  "info": {
    "title": "Sistema de Gestión de Transportes — API",
    "description": "DOC-E-BE · Contratos API definitivos (REST /api/v1). Errores application/problem+json (ADR-013). Autorización por permiso + alcance + política; el tenant se resuelve SIEMPRE desde la sesión.",
    "version": "1.0.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "CatalogDto": {
        "type": "object",
        "properties": {}
      },
      "CatalogUpdateDto": {
        "type": "object",
        "properties": {}
      },
      "ActiveDto": {
        "type": "object",
        "properties": {}
      },
...
      "AlertActionDto": {
        "type": "object",
        "properties": {}
      },
      "TelemetryBatchDto": {
        "type": "object",
        "properties": {}
      },
...
      "AdvanceDto": {
        "type": "object",
        "properties": {}
      },
...
      "PassengerDto": {
        "type": "object",
        "properties": {}
      },
      "PassengerUpdateDto": {
        "type": "object",
        "properties": {}
      },
      "CancelDto": {
        "type": "object",
        "properties": {}
      },
...
    }
  }
}